define own build backend
this removes setup.py altogether in turn preventing legacy usage
