+++ .pre-commit-config.yaml
@@ -22,6 +22,7 @@ repos:
     rev: 6.0.0
     hooks:
     -   id: flake8
+        exclude: scm_hack_build_backend.py
 -   repo: https://github.com/asottile/pyupgrade
     rev: v3.3.1
     hooks:
